web: wallet 데이터를 prop으로 넘겨줌
